use `--preemptable` appropriately when set to false
Related to https://github.com/gitpod-io/gitpod-gke-guide/pull/7

Currently, As we seem to use the `{PREEMPTABLE}` env directly,
It causes the following error when set to `false`:

```
ERROR: (gcloud.container.clusters.create) unrecognized arguments: false
```

This PR updates to use a separate variable, which is set `--preemptable`
when `PREEMPTABLE` is set to true, **empty otherwise**. Thus, preventing
`false` to mistakenly passed as an argument.

This PR also updates the usage at `gcloud.container.node-pools.create`
to not pass empty argument when the value is `false`.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>